First successful routing. Function calls crash though
